spikes: consistent deployment scripts
