+++ code/spikes/dev.000/script/deploy.mts
@@ -1,37 +1,21 @@
 #!/usr/bin/env ts-node
-import { Filesystem, NodeFs } from 'sys.fs.node';
-import { Content } from 'sys.pkg';
-import { Text } from 'sys.text/node';
-import { rx } from 'sys.util';
-
-import { Pkg } from '../src/index.pkg.mjs';
+import { bundle, bus } from './bundle.mjs';
 import { pushToVercel } from './deploy.vercel.mjs';
 
-const bus = rx.bus();
-
-const dir = async (dir: string) => {
-  const store = await Filesystem.client(NodeFs.resolve(dir), { bus });
-  return store.fs;
-};
-
-const targetdir = await dir('./dist.cell/');
-
-const bundler = await Content.bundler({
-  Text,
-  sources: {
-    app: await dir('./dist/web'),
-    src: await dir('./src/'),
-    data: await dir('./src.data'),
-    log: await dir('./dist.cell/.log/'),
-  },
-});
-
-const version = Pkg.version;
-const bundle = await bundler.write.bundle(targetdir, version);
+/**
+ * Deploy
+ */
 const deployment = await pushToVercel({
-  version,
+  bus,
+  version: bundle.version,
   fs: bundle.fs,
   source: bundle.dirs.app,
 });
 
-bundler.logger.write({ bundle, deployment });
+console.log('-------------------------------------------');
+console.log('deployed', deployment.status);
+
+/**
+ * Log results.
+ */
+await bundle.logger.write({ bundle, deployment });
