+++ code/spikes/dev.000/src/index.pkg.mts
@@ -11,7 +11,7 @@
 
 export const Pkg: ModuleDef = {
   name: 'spike.dev',
-  version: '0.0.14',
+  version: '0.0.15',
   dependencies: {
     'sys.types': '0.0.0',
     'sys.util': '0.0.0',
