--- code/spikes/identity.000/script/deploy.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env
-
-yarn audit 
-yarn version --patch --no-git-tag-version
-yarn prepare
-ts-node-esm -T ./script/deploy.mts
