+++ code/spikes/identity.000/src/index.pkg.mts
@@ -11,7 +11,7 @@
 
 export const Pkg: ModuleDef = {
   name: 'spike.identity',
-  version: '0.0.6',
+  version: '0.0.7',
   dependencies: {
     'sys.types': '0.0.0',
     'sys.util': '0.0.0',
